requirements.txt
notion-client>=2.2.1
python-telegram-bot>=20.0
groq>=0.4.0
python-dotenv>=1.0.0
duckduckgo-search>=4.0.0